requirements.txt
telethon==1.34.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
python-telegram-bot==20.8
openai==1.12.0
